Addressed comments, combined tests
