+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/broker/msa/TestCase2637853.kt
@@ -70,7 +70,6 @@ class TestCase2637853 : AbstractMsaBrokerTest() {
                 .handlePrompt(username, password)
         }, TokenRequestTimeout.MEDIUM)
         authResult.assertSuccess()
-        verifyAccountInAndroidSettings(mLabAccount)
 
         // SECOND REQUEST WITHOUT LOGIN HINT
         val noLoginHintParams = MsalAuthTestParams.builder()
